pd_login/requirements.txt
django>=2.2
djangorestframework
django-rest-swagger
django-webpack-loader
django-cors-headers
coverage
unipath
requests
pycryptodome
